+++ expected_returns/backend/main.py
@@ -1,3 +1,6 @@
+import json
+
+import requests
 import uvicorn
 from fastapi import FastAPI
 
@@ -20,5 +23,15 @@ async def read_cape(scope: str):
     return result
 
 
+@app.post("/update/{scope}")
+async def update_cape(scope: str):
+    result = cape_ratio(scope)
+    try:
+        r = requests.post(url=f"http://database:8080/{scope}", data=json.dumps(result))
+    except Exception as e:
+        r = requests.post(url=f"http://127.0.0.1:8080/{scope}", data=json.dumps(result))
+    return r.json()
+
+
 if __name__ == "__main__":
-    uvicorn.run("main:app", host="0.0.0.0", port=8000)
+    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)